requirements.txt
amqp==5.3.1
aniso8601==9.0.1
attrs==24.2.0
billiard==4.2.1
blinker==1.9.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distro==1.9.0
Flask==3.1.0
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jpype1==1.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
MarkupSafe==3.0.2
numpy==2.0.2
packaging==24.2
pandas==2.2.3
prompt_toolkit==3.0.48
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.1
referencing==0.35.1
rpds-py==0.22.3
six==1.17.0
SQLAlchemy==2.0.36
tabula-py==2.10.0
typing_extensions==4.12.2
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
